web-dashboard/backend/requirements.txt
# Core FastAPI and async support
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
email-validator>=2.1.0.post1

# Authentication and security
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
bcrypt==4.0.1

# Database and caching
redis>=5.0.1
aiosqlite>=0.20.0
sqlalchemy>=2.0.23
alembic>=1.13.1

# HTTP client and utilities
httpx>=0.27.0
aiofiles>=23.2.1

# Data processing and monitoring
psutil>=5.9.6
numpy>=1.26.2
pandas>=2.2.0

# WebSocket support
websockets>=12.0

# Logging and monitoring
structlog>=23.2.0
sentry-sdk[fastapi]>=2.7.0
prometheus-fastapi-instrumentator>=6.0.0

# Additional utilities
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
